requirements.txt
streamlit
requests
fastapi
uvicorn
python-multipart
python-dotenv
langchain-core
langchain-community
langchain-groq
langchain-openai
langchain-google-genai
langchain-text-splitters
langchain-huggingface
langchain-chroma
chromadb
pypdf
sentence-transformers